* commit '0e047a048541755228e07611dfdb44cb50a7d699':
  SS-3902,Refactor PEER environmnet PEK creation

+++ management/server/core/security-manager/security-manager-api/src/main/java/io/subutai/core/security/api/dao/SecurityDataService.java
@@ -17,7 +17,7 @@ public interface SecurityDataService
     /******************************************
      * Store Public key in the DB
      */
-    public void saveKeyIdentityData( String hostId ,String sKeyId,String pKeyId, short type );
+    public void saveKeyIdentityData( String hostId ,String sKeyId,String pKeyId, int type );
 
 
     /******************************************
@@ -41,7 +41,7 @@ public interface SecurityDataService
     /******************************************
      *
      */
-    void saveSecretKeyData( String fingerprint, byte[] data, String pwd, short type );
+    void saveSecretKeyData( String fingerprint, byte[] data, String pwd, int type );
 
 
     /******************************************
@@ -50,15 +50,36 @@ public interface SecurityDataService
     void removeSecretKeyData( String fingerprint );
 
 
+    /******************************************
+     *
+     */
     void saveKeyTrustData( String sourceId, String targetId, int trustLevel );
 
+    /******************************************
+     *
+     */
     void removeKeyTrustData( long id );
 
+    /******************************************
+     *
+     */
     void removeKeyTrustData( String sourceId );
 
+
+    /******************************************
+     *
+     */
     void removeKeyTrustData( String sourceId, String targetId );
 
+
+    /******************************************
+     *
+     */
     SecurityKeyTrust getKeyTrustData( long id );
 
+
+    /******************************************
+     *
+     */
     List<SecurityKeyTrust> getKeyTrustData( String sourceId );
 }
